refactored bot code to use cogs for commands
