+++ constants/environment.py
@@ -1,0 +1,5 @@
+from utilities.environment import get_env_var
+
+DATABASE_URL = get_env_var('DATABASE_URL')
+DISCORD_TOKEN = get_env_var('DISCORD_TOKEN')
+SECRET_GUILD_ID = get_env_var('DATABASE_URL')
